split generic 'utils' and 'canadata'
