+++ src/canadata/gameState.ts
