+++ src/canadata/settings.ts
