+++ src/test/canadata/gameState.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
-import defaultGameState from "../../utils/gameState.ts";
-import { potCodes } from "../../utils/dataBank.ts";
+import defaultGameState from "../../canadata/gameState.ts";
+import { potCodes } from "../../canadata/dataBank.ts";
 
 describe("verify the gameState", () => {
   it("defaultNewGameState should have a potCode from the dataBank", () => {
